Id security implementation.

+++ parking_lot/src/main/java/se/iths/parking_lot/controllers/ThymeleafUserController.java
@@ -32,41 +32,65 @@ public class ThymeleafUserController {
     }
 
     @GetMapping("{id}")
-    public String userById(@PathVariable("id") Long id, Model model) throws UserNotFoundException {
-        User user = userService.getById(id);
-        List<ParkingLot> parkingLots = parkingLotService.getAll();
-        AddUserToQueueDto addUserToQueueDto = new AddUserToQueueDto(-1L, false);
-        model.addAttribute("user", user);
-        model.addAttribute("parkingLots", parkingLots);
-        model.addAttribute("addUserToQueueDto", addUserToQueueDto);
-        return "user_by_id";
+    public String userById(@CookieValue("token") String token, @PathVariable("id") Long id, Model model) {
+        User currentUser = userService.getUserByToken(token);
+        if (id.equals(currentUser.getId())) {
+            List<ParkingLot> parkingLots = parkingLotService.getAll();
+            AddUserToQueueDto addUserToQueueDto = new AddUserToQueueDto(-1L, false);
+            model.addAttribute("user", currentUser);
+            model.addAttribute("parkingLots", parkingLots);
+            model.addAttribute("addUserToQueueDto", addUserToQueueDto);
+            return "user_by_id";
+        } else {
+            return "redirect:/redirect";
+        }
     }
 
     @GetMapping("{userId}/queueSlots/{slotId}/remove")
-    public String removeQueueSlots(@PathVariable("userId") Long userId, @PathVariable("slotId") Long slotId, Model model) {
-        userService.removeQueueSlot(slotId);
-        return "redirect:/tl_users/{userId}";
+    public String removeQueueSlots(@CookieValue("token") String token, @PathVariable("userId") Long userId, @PathVariable("slotId") Long slotId, Model model) {
+        User currentUser = userService.getUserByToken(token);
+        if (userId.equals(currentUser.getId())) {
+            userService.removeQueueSlot(slotId);
+            return "redirect:/tl_users/{userId}";
+        } else {
+            return "redirect:/redirect";
+        }
     }
 
     @GetMapping("{userId}/parkingSlots/{slotId}/remove")
-    public String removeParkingSlots(@PathVariable("userId") Long userId, @PathVariable("slotId") Long slotId, Model model) throws ParkingSlotNotFoundException {
-        parkingSlotService.removeUserFromParkingSlot(slotId);
-        return "redirect:/tl_users/{userId}";
+    public String removeParkingSlots(@CookieValue("token") String token, @PathVariable("userId") Long userId, @PathVariable("slotId") Long slotId, Model model) throws ParkingSlotNotFoundException {
+        User currentUser = userService.getUserByToken(token);
+        if (userId.equals(currentUser.getId())) {
+            parkingSlotService.removeUserFromParkingSlot(slotId);
+            return "redirect:/tl_users/{userId}";
+        } else {
+            return "redirect:/redirect";
+        }
     }
 
     @PostMapping("{userId}/queue")
-    public String addUserToQueue(@PathVariable("userId") Long userId, @ModelAttribute AddUserToQueueDto addUserToQueueDto, Model model) throws UserNotFoundException, ParkingLotNotFoundException {
-        model.addAttribute("dto", addUserToQueueDto);
-        userService.queryToParkingLot(userId, addUserToQueueDto.parkingLotId(), addUserToQueueDto.electricCharge());
-        return "redirect:/tl_users/{userId}";
+    public String addUserToQueue(@CookieValue("token") String token, @PathVariable("userId") Long userId, @ModelAttribute AddUserToQueueDto addUserToQueueDto, Model model) throws UserNotFoundException, ParkingLotNotFoundException {
+        User currentUser = userService.getUserByToken(token);
+        if (userId.equals(currentUser.getId())) {
+            model.addAttribute("dto", addUserToQueueDto);
+            userService.queryToParkingLot(userId, addUserToQueueDto.parkingLotId(), addUserToQueueDto.electricCharge());
+            return "redirect:/tl_users/{userId}";
+        } else {
+            return "redirect:/redirect";
+        }
     }
 
     @GetMapping("{userId}/edit")
-    public String editUser(@PathVariable("userId") Long userId, @RequestParam(name = "constrainException", defaultValue = "false") Boolean constrainException, Model model) throws UserNotFoundException {
-        User user = userService.getById(userId);
-        model.addAttribute("user", user);
-        model.addAttribute("constrainException", constrainException);
-        return "user_edit";
+    public String editUser(@CookieValue("token") String token, @PathVariable("userId") Long userId, @RequestParam(name = "constrainException", defaultValue = "false") Boolean constrainException, Model model) throws UserNotFoundException {
+        User currentUser = userService.getUserByToken(token);
+        if (userId.equals(currentUser.getId())) {
+            User user = userService.getById(userId);
+            model.addAttribute("user", user);
+            model.addAttribute("constrainException", constrainException);
+            return "user_edit";
+        } else {
+            return "redirect:/redirect";
+        }
     }
 
     @PostMapping("{userId}/edit/submit")
